relocate section data
